Fix tests (added ordering for blinking tests)
